# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Niclas-REST-API",
    "description": "BOOTCAMP REST API NICLAS GREGOR",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/book": {
      "get": {
        "summary": "Returns a list of books",
        "description": "Overview of all books including their author, year of publication, title and ID",
        "responses": {
          "200": {
            "description": "A JSON array of books with their information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adds new book to database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "year_written": {
                    "type": "integer"
                  },
                  "author": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nothing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/author": {
      "get": {
        "summary": "Returns a list of all authors",
        "responses": {
          "200": {
            "description": "A JSON array of all authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/title": {
      "get": {
        "summary": "Returns a list of all titles",
        "responses": {
          "200": {
            "description": "A JSON array of all titles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/{bookId}": {
      "get": {
        "summary": "Returns a book by ID",
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of the book with the matching ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deletes a book by ID",
        "parameters": [
          {
            "in": "path",
            "name": "bookId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of the deleted book",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/year/{yearOfPublication}": {
      "get": {
        "summary": "Returns a list of book written in the matching year",
        "parameters": [
          {
            "in": "path",
            "name": "yearOfPublication",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int16",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of the books written in the matching year",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/author/{author}": {
      "get": {
        "summary": "Returns a list of books written by the matching author",
        "parameters": [
          {
            "in": "path",
            "name": "author",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of the books written by the matching author",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/book/title/{title}": {
      "get": {
        "summary": "Returns a book by title",
        "parameters": [
          {
            "in": "path",
            "name": "title",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON array of the book matching the title",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "yearOfPublication": {
                      "type": "integer"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}